+++ ft-logic/io/src/lib.rs
@@ -1,8 +1,7 @@
 #![no_std]
-use ft_main_io::LogicAction;
 use gmeta::{In, InOut, Metadata};
 use gstd::{prelude::*, ActorId};
-use primitive_types::H256;
+use primitive_types::{H256, H512};
 pub struct FLogicMetadata;
 pub mod instruction;
 use instruction::Instruction;
@@ -45,6 +44,34 @@ pub enum FTLogicAction {
     Clear(H256),
     UpdateStorageCodeHash(H256),
     MigrateStorages,
+}
+
+#[derive(Encode, Debug, Decode, TypeInfo, Copy, Clone)]
+pub enum LogicAction {
+    Mint {
+        recipient: ActorId,
+        amount: u128,
+    },
+    Burn {
+        sender: ActorId,
+        amount: u128,
+    },
+    Transfer {
+        sender: ActorId,
+        recipient: ActorId,
+        amount: u128,
+    },
+    Approve {
+        approved_account: ActorId,
+        amount: u128,
+    },
+    Permit {
+        owner_account: ActorId,
+        approved_account: ActorId,
+        amount: u128,
+        permit_id: u128,
+        sign: H512,
+    },
 }
 
 #[derive(Encode, Decode, TypeInfo)]
