+++ ft-main/io/src/lib.rs
@@ -24,7 +24,7 @@ pub struct FTokenState {
 pub enum FTokenAction {
     Message {
         transaction_id: u64,
-        payload: LogicAction,
+        payload: Vec<u8>,
     },
     UpdateLogicContract {
         ft_logic_code_hash: H256,
@@ -34,47 +34,6 @@ pub enum FTokenAction {
     GetPermitId(ActorId),
     Clear(H256),
     MigrateStorageAddresses,
-}
-
-#[derive(Encode, Decode, TypeInfo, Debug)]
-pub enum FTokenInnerAction {
-    Message(Vec<u8>),
-    UpdateLogicContract {
-        ft_logic_code_hash: H256,
-        storage_code_hash: H256,
-    },
-    GetBalance(ActorId),
-    GetPermitId(ActorId),
-    Clear(H256),
-    MigrateStorageAddresses,
-}
-
-#[derive(Encode, Debug, Decode, TypeInfo, Copy, Clone)]
-pub enum LogicAction {
-    Mint {
-        recipient: ActorId,
-        amount: u128,
-    },
-    Burn {
-        sender: ActorId,
-        amount: u128,
-    },
-    Transfer {
-        sender: ActorId,
-        recipient: ActorId,
-        amount: u128,
-    },
-    Approve {
-        approved_account: ActorId,
-        amount: u128,
-    },
-    Permit {
-        owner_account: ActorId,
-        approved_account: ActorId,
-        amount: u128,
-        permit_id: u128,
-        sign: H512,
-    },
 }
 
 #[derive(Debug, Encode, Decode, TypeInfo, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
