+++ ft-main/tests/high_load_tests.rs
@@ -29,6 +29,7 @@ fn high_load_mint() {
     }
 }
 
+#[ignore]
 #[test]
 fn high_load_transfer() {
     const FIRST_ID: u64 = 100;
