+++ ft-main/tests/utils.rs
@@ -1,4 +1,5 @@
 use ft_main_io::*;
+use ft_logic_io::{LogicAction};
 use gstd::{prelude::*, ActorId};
 use gtest::{Log, Program, System};
 use sp_core::sr25519::Signature;
@@ -70,7 +71,7 @@ impl FToken for Program<'_> {
         let payload = LogicAction::Mint {
             recipient: account.into(),
             amount,
-        };
+        }.encode();
         self.send_message_and_check_res(
             from,
             FTokenAction::Message {
@@ -85,7 +86,7 @@ impl FToken for Program<'_> {
         let payload = LogicAction::Burn {
             sender: account.into(),
             amount,
-        };
+        }.encode();
         self.send_message_and_check_res(
             from,
             FTokenAction::Message {
@@ -109,7 +110,7 @@ impl FToken for Program<'_> {
             sender: sender.into(),
             recipient: recipient.into(),
             amount,
-        };
+        }.encode();
         self.send_message_and_check_res(
             from,
             FTokenAction::Message {
@@ -131,7 +132,7 @@ impl FToken for Program<'_> {
         let payload = LogicAction::Approve {
             approved_account: approved_account.into(),
             amount,
-        };
+        }.encode();
         self.send_message_and_check_res(
             from,
             FTokenAction::Message {
@@ -159,7 +160,7 @@ impl FToken for Program<'_> {
             amount,
             permit_id,
             sign: sign.into(),
-        };
+        }.encode();
         self.send_message_and_check_res(
             from,
             FTokenAction::Message {
